requirements.txt
Flask==3.0.3
tensorflow==2.19.0
Pillow==10.2.0
numpy==1.26.4
Werkzeug==3.0.3
gunicorn==22.0.0